Separate OfType into its own files
Only moved the code; didn't change it at all.
