+++ src/libraries/System.Linq/src/System/Linq/OfType.cs
@@ -9,45 +9,30 @@ namespace System.Linq
 {
     public static partial class Enumerable
     {
-        public static IEnumerable<
-#nullable disable // there's no way to annotate the connection of the nullability of TResult to that of the source
-                TResult
-#nullable restore
-                > Cast<TResult>(this IEnumerable source)
+        public static IEnumerable<TResult> OfType<TResult>(this IEnumerable source)
         {
-            if (source is IEnumerable<TResult> typedSource)
-            {
-                return typedSource;
-            }
-
             if (source == null)
             {
                 ThrowHelper.ThrowArgumentNullException(ExceptionArgument.source);
             }
 
-            if (source is ICollection collection)
+            if (default(TResult) is not null && source is IEnumerable<TResult> typedSource)
             {
-                return new CastICollectionIterator<TResult>(collection);
+                // The source was already an IEnumerable<TResult> and TResult can't be null. As
+                // such, all values the original input can yield are valid, and we can just return
+                // the strongly-typed input directly as if this were Cast rather than OfType.
+                return typedSource;
             }
 
-            return CastIterator<TResult>(source);
+            return new OfTypeIterator<TResult>(source);
         }
 
-        private static IEnumerable<TResult> CastIterator<TResult>(IEnumerable source)
+        private sealed partial class OfTypeIterator<TResult>(IEnumerable source) : Iterator<TResult>
         {
-            foreach (object obj in source)
-            {
-                yield return (TResult)obj;
-            }
-        }
-
-        [DebuggerDisplay("Count = {Count}")]
-        private sealed partial class CastICollectionIterator<TResult>(ICollection source) : Iterator<TResult>
-        {
-            private readonly ICollection _source = source;
+            private readonly IEnumerable _source = source;
             private IEnumerator? _enumerator;
 
-            public override Iterator<TResult> Clone() => new CastICollectionIterator<TResult>(_source);
+            public override Iterator<TResult> Clone() => new OfTypeIterator<TResult>(_source);
 
             public override bool MoveNext()
             {
@@ -60,10 +45,13 @@ namespace System.Linq
 
                     case 2:
                         Debug.Assert(_enumerator != null);
-                        if (_enumerator.MoveNext())
+                        while (_enumerator.MoveNext())
                         {
-                            _current = (TResult)_enumerator.Current;
-                            return true;
+                            if (_enumerator.Current is TResult result)
+                            {
+                                _current = result;
+                                return true;
+                            }
                         }
 
                         Dispose();
